Added live load cell graph
